+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to the "uihelper" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [Unreleased 0.0.10]
+- html2scss support more filetype('html','htm','jsp','php','vue','jsx','tsx','wxml','javascriptreact','typescriptreact')
 ## [Unreleased 0.0.4]
 - fix install in windows os
 ## [Unreleased 0.0.3]